Plot alignment scatterplot

Plot alignment scatterplot

Usage

plot_scatter_alignment_exposure(
  data,
  floor_outliers,
  cap_outliers,
  group_var,
  currency
)

Arguments

data

data.frame. Should have the same format as output of prep_scatter() and contain columns: 'name', 'buildout', 'phaseout', 'net', and any column implied by group_var.

floor_outliers

Numeric. Floor which should be applied to the alignment values in the data. Values smaller than floor are plotted on the border of the plot.

cap_outliers

Numeric. Cap which should be applied to the alignment values in the data. Values bigger than cap are plotted on the border of the plot.

group_var

Character. Character specifying the variable that contains the groups by which to analyse the loan books.

currency

Character. Currency to display in the plot labels.

Value

object of type "ggplot"

Examples

# TODO